redacted/requirements.txt
reportlab==4.1.0
pillow==10.2.0
Flask==3.0.0